Decompose bookSyncService into parser, diff, and orchestrator
BookSyncService was a god object with three unrelated responsibilities in one
440-line class: a ~140-line MediaWiki award-table parser, a ~96-line Notion diff
builder, and the sync orchestration — parsing logic that duplicated wiki.parser's
concern, diff logic that belonged with the other pure comparators, and I/O
orchestration all tangled together.

Split along the seams:
- wiki.parser.ts gains a pure `parseAwardTable(wikitext, awardName): Book[]` —
  the entire table-parsing body, now unit-testable without a service or mock.
- services/bookDiff.ts holds the pure payload builders: buildAuthorTags (moved
  here), buildBookUpdates (was compareBooks), and buildNewBookProperties
  (extracted from the new-row branch, collapsing the duplicated award/"Wszystkie"
  logic into one mergeAwards helper shared by both paths).
- BookSyncService keeps only orchestration (fetch → merge/dedup → write) and
  shrinks 440 → 164 lines; fetchBooksFromMediaWiki is now fetch + delegate.

Behavior is unchanged. Tests moved to bookDiff.test.ts and a parseAwardTable
test added to wiki.parser.test.ts; the compare/author cases are byte-equivalent.
137 passing, lint clean, build OK.
