+++ services/__tests__/bookSyncService.test.ts
@@ -1,36 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { BookSyncService, buildAuthorTags } from '../bookSyncService';
+import { BookSyncService } from '../bookSyncService';
 import { NotionAdapter } from '../../notion.adapter';
 import { WikiAdapter } from '../../wiki.adapter';
-import { Book, NotionBook } from '../../src/types';
-
-describe('buildAuthorTags', () => {
-  it('never emits a multi_select tag containing a comma, even when a mapping injects one', () => {
-    // "Liu Cixin" maps to "Liu Cixin, Ken Liu"; Notion rejects commas in options.
-    const tags = buildAuthorTags('Liu Cixin');
-    expect(tags).toEqual(['Liu Cixin', 'Ken Liu']);
-    expect(tags.some(t => t.includes(','))).toBe(false);
-  });
-
-  it('splits an already-merged comma author string into clean tags and dedups', () => {
-    const tags = buildAuthorTags('Liu Cixin, Ken Liu');
-    expect(tags).toEqual(['Liu Cixin', 'Ken Liu']);
-  });
-
-  it('handles the Hamilton pseudonym mapping without a comma tag', () => {
-    const tags = buildAuthorTags('Edmond Hamilton (jako Brett Sterling)');
-    expect(tags).toEqual(['Edmond Hamilton', 'Brett Sterling']);
-    expect(tags.some(t => t.includes(','))).toBe(false);
-  });
-
-  it('passes ordinary single authors through unchanged', () => {
-    expect(buildAuthorTags('Stanisław Lem')).toEqual(['Stanisław Lem']);
-  });
-
-  it('returns [] for empty input', () => {
-    expect(buildAuthorTags('')).toEqual([]);
-  });
-});
 
 describe('BookSyncService', () => {
   let service: BookSyncService;
@@ -79,198 +50,6 @@ describe('BookSyncService', () => {
         polishTitle: 'Solaris',
         award: 'Nagroda Test'
       }));
-    });
-  });
-
-  describe('compareBooks', () => {
-    it('returns updates when titles differ', () => {
-      const existing: NotionBook = {
-        id: '1',
-        plTitle: 'Stary',
-        origTitle: '', // Empty to trigger update
-        author: 'Lem',
-        year: '2000',
-        awards: [],
-        zrodlo: [],
-        currentWydawnictwo: '',
-        currentSeria: '',
-        currentCzesccyklu: false,
-        lp: '1',
-        plTitleRichText: [],
-        origTitleRichText: []
-      };
-      const newBook: Book = {
-        polishTitle: 'Nowy',
-        originalTitle: 'New',
-        author: 'Lem',
-        year: '2000',
-        award: 'Test',
-        polishTitleLink: null
-      };
-
-      const updates = service.compareBooks(existing, newBook);
-
-      expect(updates).toHaveProperty('Tytuł polski');
-      expect(updates).toHaveProperty('Tytuł oryginalny');
-    });
-
-    it('returns empty object when books are identical', () => {
-      const existing: NotionBook = {
-        id: '1',
-        plTitle: 'Solaris',
-        origTitle: 'Solaris',
-        author: 'Stanisław Lem',
-        year: '1961',
-        awards: ['Nagroda Test'],
-        zrodlo: [],
-        currentWydawnictwo: '',
-        currentSeria: '',
-        currentCzesccyklu: false,
-        lp: '1',
-        plTitleRichText: [],
-        origTitleRichText: []
-      };
-      const newBook: Book = {
-        polishTitle: 'Solaris',
-        originalTitle: 'Solaris',
-        author: 'Stanisław Lem',
-        year: '1961',
-        award: 'Nagroda Test',
-        awards: ['Nagroda Test'],
-        polishTitleLink: null
-      };
-
-      const updates = service.compareBooks(existing, newBook);
-
-      expect(updates).toEqual({});
-    });
-
-    it('merges authors instead of replacing manually added ones', () => {
-      const existing: NotionBook = {
-        id: '1',
-        plTitle: 'Diuna',
-        origTitle: 'Dune',
-        author: 'Frank Herbert, Brian Herbert',
-        year: '1965',
-        awards: ['Nagroda Hugo'],
-        zrodlo: [],
-        currentWydawnictwo: '',
-        currentSeria: '',
-        currentCzesccyklu: false,
-        lp: '1',
-        plTitleRichText: [],
-        origTitleRichText: []
-      };
-      const newBook: Book = {
-        polishTitle: 'Diuna',
-        originalTitle: 'Dune',
-        author: 'Frank Herbert',
-        year: '1965',
-        award: 'Nagroda Hugo',
-        awards: ['Nagroda Hugo'],
-        polishTitleLink: null
-      };
-
-      const updates = service.compareBooks(existing, newBook);
-
-      // Wiki has fewer authors than Notion — nothing to add, nothing removed
-      expect(updates).not.toHaveProperty('Autor');
-    });
-
-    it('adds new wiki authors while keeping existing ones', () => {
-      const existing: NotionBook = {
-        id: '1',
-        plTitle: 'Diuna',
-        origTitle: 'Dune',
-        author: 'Frank Herbert',
-        year: '1965',
-        awards: ['Nagroda Hugo'],
-        zrodlo: [],
-        currentWydawnictwo: '',
-        currentSeria: '',
-        currentCzesccyklu: false,
-        lp: '1',
-        plTitleRichText: [],
-        origTitleRichText: []
-      };
-      const newBook: Book = {
-        polishTitle: 'Diuna',
-        originalTitle: 'Dune',
-        author: 'Frank Herbert, Kevin J. Anderson',
-        year: '1965',
-        award: 'Nagroda Hugo',
-        awards: ['Nagroda Hugo'],
-        polishTitleLink: null
-      };
-
-      const updates = service.compareBooks(existing, newBook);
-
-      expect(updates).toHaveProperty('Autor');
-      const names = updates['Autor'].multi_select.map((o: any) => o.name);
-      expect(names).toContain('Frank Herbert');
-      expect(names).toContain('Kevin J. Anderson');
-    });
-
-    it('is idempotent when Notion stores a different-case author (no churny re-updates)', () => {
-      // Notion keeps its own option casing ("van Vogt"); wiki normalizes to
-      // "Van Vogt". These are the same author — must NOT trigger an update.
-      const existing: NotionBook = {
-        id: '1',
-        plTitle: 'Slan',
-        origTitle: 'Slan',
-        author: 'A. E. van Vogt', // stored casing from Notion
-        year: '1941',
-        awards: ['Nagroda Hugo'],
-        zrodlo: [],
-        currentWydawnictwo: '',
-        currentSeria: '',
-        currentCzesccyklu: false,
-        lp: '1',
-        plTitleRichText: [],
-        origTitleRichText: []
-      };
-      const newBook: Book = {
-        polishTitle: 'Slan',
-        originalTitle: 'Slan',
-        author: 'A. E. Van Vogt', // wiki-normalized casing
-        year: '1941',
-        award: 'Nagroda Hugo',
-        awards: ['Nagroda Hugo'],
-        polishTitleLink: null
-      };
-
-      const updates = service.compareBooks(existing, newBook);
-      expect(updates).not.toHaveProperty('Autor');
-    });
-
-    it('is idempotent for a multi-author book with mixed casing', () => {
-      const existing: NotionBook = {
-        id: '1',
-        plTitle: 'The Winged Man',
-        origTitle: 'The Winged Man',
-        author: 'A. E. van Vogt, Edna Mayne Hull',
-        year: '1945',
-        awards: ['Nagroda Hugo'],
-        zrodlo: [],
-        currentWydawnictwo: '',
-        currentSeria: '',
-        currentCzesccyklu: false,
-        lp: '1',
-        plTitleRichText: [],
-        origTitleRichText: []
-      };
-      const newBook: Book = {
-        polishTitle: 'The Winged Man',
-        originalTitle: 'The Winged Man',
-        author: 'A. E. Van Vogt, Edna Mayne Hull',
-        year: '1945',
-        award: 'Nagroda Hugo',
-        awards: ['Nagroda Hugo'],
-        polishTitleLink: null
-      };
-
-      const updates = service.compareBooks(existing, newBook);
-      expect(updates).not.toHaveProperty('Autor');
     });
   });
