+++ services/bookSyncService.ts
@@ -1,39 +1,14 @@
 import pLimit from "p-limit";
 import { NotionAdapter } from "../notion.adapter";
 import { WikiAdapter } from "../wiki.adapter";
-import { calculateSimilarity, countCommonWords, cleanTitle, sanitizeNotionString, sanitizeNotionTag, isValidUrl } from "../utils";
+import { WikiParser } from "../wiki.parser";
+import { calculateSimilarity, countCommonWords, cleanTitle } from "../utils";
 import { Book, NotionBook, SyncEvent, SyncParams } from "../src/types";
 import { normalizeData } from "./dataNormalizer";
+import { buildBookUpdates, buildNewBookProperties } from "./bookDiff";
 import { createLogger } from "../logger";
 
 const log = createLogger("BookSync");
-
-/**
- * Buduje listę tagów autora dla pola multi_select. KOLEJNOŚĆ MA ZNACZENIE:
- * najpierw normalizacja (mapowania potrafią rozwinąć jedno nazwisko w kilka
- * rozdzielonych przecinkiem, np. "Liu Cixin" → "Liu Cixin, Ken Liu"), potem
- * podział po przecinku, a sanitizacja NA KOŃCU. `sanitizeNotionTag` usuwa
- * przecinki (Notion nie dopuszcza ich w opcjach select/multi_select) — gdy
- * sanitizacja biegła PRZED normalizacją, wstrzyknięty przez mapowanie przecinek
- * trafiał do nazwy opcji i Notion odrzucał cały wiersz (laureat gubiony w ciszy).
- * Deduplikacja jest case-insensitive, z zachowaniem pierwszej napotkanej pisowni.
- */
-export function buildAuthorTags(raw: string): string[] {
-  if (!raw) return [];
-  const seen = new Set<string>();
-  const tags: string[] = [];
-  for (const part of raw.split(",")) {
-    const normalized = normalizeData(part.trim(), "author");
-    for (const piece of normalized.split(",")) {
-      const tag = sanitizeNotionTag(piece);
-      if (tag && !seen.has(tag.toLowerCase())) {
-        seen.add(tag.toLowerCase());
-        tags.push(tag);
-      }
-    }
-  }
-  return tags;
-}
 
 export class BookSyncService {
   constructor(private notion: NotionAdapter, private wiki: WikiAdapter) {}
@@ -48,240 +23,9 @@ export class BookSyncService {
     }
 
     sendEvent({ type: "status", message: `Dekodowanie tablic wyników: ${awardName}...` });
-    const books: Book[] = [];
-    
-    let processedWikitext = wikitext;
-    if (awardName === "Nagroda Locus") {
-      processedWikitext = processedWikitext.replace(/={2,}\s*Zwycięzcy w kategorii Powieść dla młodzieży\s*={2,}[^]*?(?=\n==|$)/i, '');
-    }
-
-    let cleanedWiki = processedWikitext
-      .replace(/\{\{sortname\|([^|}]+)\|([^|}]+)(?:\|[^}]+)?\}\}/gi, '$1 $2')
-      .replace(/<br\s*\/?>/gi, ' ')
-      .replace(/<[^>]+>/g, '')
-      .replace(/\|\}[\s\S]*?\{\|[^\n]*\n/g, '\n|-\n');
-
-    const rows = cleanedWiki.split(/\|-/);
-    let lastYear = "";
-    let colMap = { rok: 0, autor: 1, tytulOryginalny: 2, tytulPolski: 3, expectedLength: 4 };
-
-    const parseCell = (cell: string, extractLink = false, extractItalics = false) => {
-      if (!cell) return extractLink ? { text: "", link: null } : "";
-      let text = cell.trim();
-      let link = null;
-      
-      if (extractLink) {
-        const linkMatch = text.match(/\[\[([^\]|]+)(?:\|([^\]]+))?\]\]/);
-        if (linkMatch) {
-          const pageName = linkMatch[1].trim().replace(/ /g, '_');
-          link = `https://encyklopediafantastyki.pl/index.php?title=${encodeURIComponent(pageName)}`;
-        }
-      }
-      
-      if (extractItalics && !text.toLowerCase().includes("(aka")) {
-        const italicsMatch = text.match(/''(.+?)''/);
-        if (italicsMatch) text = italicsMatch[1];
-      }
-      
-      text = text.replace(/\[\[[^\]|]+\|([^\]]+)\]\]/g, '$1')
-                 .replace(/\[\[([^\]]+)\]\]/g, '$1')
-                 .replace(/''+/g, '')
-                 .trim();
-                 
-      return extractLink ? { text, link } : text;
-    };
-
-    for (let i = 0; i < rows.length; i++) {
-      let trimmedRow = rows[i].trim();
-      const endTableIndex = trimmedRow.indexOf('|}');
-      if (endTableIndex !== -1) trimmedRow = trimmedRow.substring(0, endTableIndex).trim();
-      if (!trimmedRow || trimmedRow.startsWith('}')) continue;
-
-      const isWinner = trimmedRow.toLowerCase().includes('background') || trimmedRow.toLowerCase().includes('bgcolor');
-      const baseName = awardName.replace('Nagroda ', '').replace('Nominacja ', '');
-      const currentAward = isWinner ? `Nagroda ${baseName}` : `Nominacja ${baseName}`;
-
-      let rowContent = trimmedRow;
-      if (!rowContent.startsWith('|') && !rowContent.startsWith('!')) rowContent = rowContent.replace(/^.*?\n/, '');
-      rowContent = '\n' + rowContent;
-      
-      const cellsRaw = rowContent.split(/\n\||\n!|\|\|/);
-      const cells = cellsRaw.map(c => {
-        const pipeIndex = c.indexOf('|');
-        if (pipeIndex !== -1 && c.substring(0, pipeIndex).includes('=')) return c.substring(pipeIndex + 1).trim();
-        return c.trim().replace(/^!/, '').replace(/^\|/, '').trim();
-      });
-
-      if (cells.length > 0 && cells[0] === '') cells.shift();
-      while (cells.length > 0 && cells[cells.length - 1] === '') cells.pop();
-      
-      if (cells.some(c => c.toLowerCase().includes('autor'))) {
-        colMap.rok = cells.findIndex(c => c.toLowerCase().includes('rok'));
-        colMap.autor = cells.findIndex(c => c.toLowerCase().includes('autor'));
-        colMap.tytulOryginalny = cells.findIndex(c => c.toLowerCase().includes('oryginalny'));
-        colMap.tytulPolski = cells.findIndex(c => c.toLowerCase().includes('polski'));
-        colMap.expectedLength = cells.length;
-        continue;
-      }
-
-      if (cells.length === 1) {
-        const cellText = parseCell(cells[0]) as string;
-        const yearOnly = cellText.replace(/['\[\]]/g, '').trim().match(/^(\d{4})$/);
-        if (yearOnly) {
-          // Wiersz zawierający wyłącznie rok (rowspan) — to nowy rocznik,
-          // nie dodatkowy autor poprzedniej książki
-          lastYear = yearOnly[1];
-        } else if (books.length > 0) {
-          const extraAuthor = cellText.replace(/\s*\(remis\)/gi, '').trim();
-          books[books.length - 1].author += ", " + extraAuthor;
-        }
-        continue;
-      }
-
-      if (cells.length < 2) continue;
-
-      let year = "", author = "", originalTitle = "", polishTitle = "", polishTitleLink = null;
-      let isFullRow = false;
-      if (colMap.rok !== -1) {
-         const possibleYearCell = cells[colMap.rok];
-         if (possibleYearCell) {
-             const yearMatch = possibleYearCell.match(/^['\[]*(\d{4})/);
-             if (yearMatch && !possibleYearCell.includes('{{sortname')) isFullRow = true;
-         }
-      } else {
-         isFullRow = cells.length >= colMap.expectedLength;
-      }
-
-      const omittedColumns = isFullRow ? 0 : Math.max(1, colMap.autor);
-      
-      if (isFullRow) {
-        const yearMatch = cells[colMap.rok]?.match(/^['\[]*(\d{4})/);
-        year = yearMatch ? yearMatch[1] : cells[colMap.rok];
-        lastYear = year;
-        author = parseCell(cells[colMap.autor] || "") as string;
-        originalTitle = parseCell(cells[colMap.tytulOryginalny] || "", false, true) as string;
-        const plParsed = parseCell(cells[colMap.tytulPolski] || "", true, true) as {text: string, link: string | null};
-        polishTitle = plParsed.text;
-        polishTitleLink = plParsed.link;
-      } else {
-        year = lastYear;
-        author = parseCell(cells[0] || "") as string;
-        originalTitle = parseCell(cells[colMap.tytulOryginalny - omittedColumns] || "", false, true) as string;
-        const plParsed = parseCell(cells[colMap.tytulPolski - omittedColumns] || "", true, true) as {text: string, link: string | null};
-        polishTitle = plParsed.text;
-        polishTitleLink = plParsed.link;
-      }
-
-      author = author.replace(/\s*\(remis\)/gi, '').trim();
-      if (author || polishTitle || originalTitle) {
-        books.push({ 
-          year, 
-          author, 
-          polishTitle: cleanTitle(polishTitle), 
-          originalTitle: normalizeData(cleanTitle(originalTitle), 'title'), 
-          polishTitleLink, award: currentAward 
-        });
-      }
-    }
+    const books = WikiParser.parseAwardTable(wikitext, awardName);
     log.info(`Sparsowano ${books.length} książek dla ${awardName}`, { awardName, pageTitle, wikitextLength: wikitext.length, books: books.length });
     return books;
-  }
-
-  compareBooks(existingBook: NotionBook, book: Book): any {
-    const updates: any = {};
-    
-    const cleanExistingPl = sanitizeNotionString(existingBook.plTitle);
-    const cleanExistingOrig = sanitizeNotionString(existingBook.origTitle);
-    const cleanNewPl = sanitizeNotionString(book.polishTitle);
-    const cleanNewOrig = sanitizeNotionString(book.originalTitle);
-
-    // Update Polish title if it differs from the wiki
-    if (cleanNewPl && cleanExistingPl !== cleanNewPl) {
-      const link = book.polishTitleLink;
-      updates["Tytuł polski"] = { rich_text: [{ text: { content: cleanNewPl, ...(isValidUrl(link) ? { link: { url: link } } : {}) } }] };
-    } else if (cleanExistingPl && existingBook.plTitle !== cleanExistingPl) {
-      const link = book.polishTitleLink;
-      updates["Tytuł polski"] = { rich_text: [{ text: { content: cleanExistingPl, ...(isValidUrl(link) ? { link: { url: link } } : {}) } }] };
-    }
-
-    if (cleanNewOrig && (!cleanExistingOrig || cleanExistingOrig === "")) {
-      updates["Tytuł oryginalny"] = { rich_text: [{ text: { content: cleanNewOrig } }] };
-    } else if (cleanExistingOrig && existingBook.origTitle !== cleanExistingOrig) {
-      updates["Tytuł oryginalny"] = { rich_text: [{ text: { content: cleanExistingOrig } }] };
-    }
-
-    // Update Author if it differs.
-    // Normalizuj i porównuj OBIE strony case-insensitive: Notion dopasowuje opcje
-    // multi_select bez względu na wielkość liter i zachowuje własną pisownię, więc
-    // normalizowanie tylko nowej wartości (np. "A. E. Van Vogt" vs zapisane
-    // "A. E. van Vogt") powodowało aktualizację przy każdym sync, która nigdy nie
-    // "chwytała". Aktualizuj tylko, gdy faktycznie dochodzi nowy autor.
-    const parseAuthors = (raw: string) => buildAuthorTags(raw);
-
-    const newAuthors = parseAuthors(book.author || "");
-    const existingAuthors = parseAuthors(existingBook.author || "");
-    const existingLower = new Set(existingAuthors.map(a => a.toLowerCase()));
-
-    // Merge authors (union) — never drop authors added manually in Notion
-    const combinedAuthors = [...existingAuthors];
-    for (const a of newAuthors) {
-      if (!existingLower.has(a.toLowerCase())) {
-        combinedAuthors.push(a);
-        existingLower.add(a.toLowerCase());
-      }
-    }
-    const authorsChanged = combinedAuthors.length !== existingAuthors.length;
-
-    if (authorsChanged && combinedAuthors.length > 0) {
-      updates["Autor"] = { multi_select: combinedAuthors.slice(0, 100).map(name => ({ name })) };
-    }
-    
-    const newAwards = (book.awards || []).map(sanitizeNotionTag).filter(Boolean);
-    const existingAwards = (existingBook.awards || []).map(sanitizeNotionTag).filter(Boolean);
-    
-    // Case-insensitive union (jak autorzy/wydawca/seria). Notion dopasowuje opcje
-    // multi_select bez względu na wielkość liter i zachowuje własną pisownię, więc
-    // porównanie case-sensitive mogło re-dodać istniejący tag (np. ręcznie wpisane
-    // "nagroda hugo" obok wygenerowanego "Nagroda Hugo"). Zob. GUIDELINES §3.
-    const awardsLower = new Set(existingAwards.map(a => a.toLowerCase()));
-    const combinedAwards = [...existingAwards];
-    let awardsUpdated = false;
-
-    for (const aw of newAwards) {
-      if (!awardsLower.has(aw.toLowerCase())) {
-        combinedAwards.push(aw);
-        awardsLower.add(aw.toLowerCase());
-        awardsUpdated = true;
-      }
-    }
-
-    const hasHugo = awardsLower.has("nagroda hugo");
-    const hasNebula = awardsLower.has("nagroda nebula");
-    const hasLocus = awardsLower.has("nagroda locus");
-    if (hasHugo && hasNebula && hasLocus && !awardsLower.has("wszystkie")) {
-      combinedAwards.push("Wszystkie");
-      awardsLower.add("wszystkie");
-      awardsUpdated = true;
-    }
-
-    if (awardsUpdated) {
-      updates["Nagroda"] = { multi_select: combinedAwards.slice(0, 100).map(name => ({ name })) };
-    }
-
-    // Update Year if it differs
-    const newYear = (book.year || "").trim();
-    const existingYears = (existingBook.year || "").split(',')
-      .map((y: string) => y.trim())
-      .filter(Boolean);
-    
-    if (newYear && !existingYears.includes(newYear)) {
-      const updatedYears = Array.from(new Set([...existingYears, newYear]))
-        .sort()
-        .map(name => ({ name }));
-      updates["Rok"] = { multi_select: updatedYears };
-    }
-
-    return updates;
   }
 
   async runBookSync(params: SyncParams, sendEvent: (data: SyncEvent) => void, checkCancellation: () => boolean) {
@@ -391,7 +135,7 @@ export class BookSyncService {
           if (isDuplicateFound) return;
 
           if (existingBook) {
-            const updates = this.compareBooks(existingBook, book);
+            const updates = buildBookUpdates(existingBook, book);
             if (Object.keys(updates).length > 0) {
               await this.notion.updatePage(existingBook.id, updates);
               updated++;
@@ -401,27 +145,7 @@ export class BookSyncService {
               syncSummary.skipped.push(bookDisplayName);
             }
           } else {
-            const properties: any = {
-              "Lp": { title: [{ text: { content: sanitizeNotionString(book.polishTitle || book.originalTitle || "Nowy") } }] },
-              "Tytuł polski": { rich_text: [{ text: { content: sanitizeNotionString(book.polishTitle || ""), ...(isValidUrl(book.polishTitleLink) ? { link: { url: book.polishTitleLink } } : {}) } }] },
-              "Tytuł oryginalny": { rich_text: [{ text: { content: sanitizeNotionString(book.originalTitle || "") } }] },
-            };
-            if (book.year) {
-              properties["Rok"] = { multi_select: [{ name: book.year.toString() }] };
-            }
-            if (book.author) {
-              const authors = buildAuthorTags(book.author);
-              properties["Autor"] = { multi_select: authors.slice(0, 100).map(name => ({ name })) };
-            }
-            if (book.awards && book.awards.length > 0) {
-              const awardsSet = new Set(book.awards.map(sanitizeNotionTag).filter(Boolean));
-              const hasHugo = awardsSet.has("Nagroda Hugo");
-              const hasNebula = awardsSet.has("Nagroda Nebula");
-              const hasLocus = awardsSet.has("Nagroda Locus");
-              if (hasHugo && hasNebula && hasLocus && !awardsSet.has("Wszystkie")) awardsSet.add("Wszystkie");
-              properties["Nagroda"] = { multi_select: Array.from(awardsSet).map(name => ({ name })) };
-            }
-            await this.notion.addRow(properties);
+            await this.notion.addRow(buildNewBookProperties(book));
             synced++;
             syncSummary.added.push(bookDisplayName);
           }
